Write the final logic for calculator
The final logic is written and calculator can now perform most basic functions in a chain.
The division and multiplication are done first and then subtraction and addition.
Floating point is supported. Result of the previos calculation cannot be yet used as the first value.
Paranthesies and some mathematical operations are not yet implemented